+++ esp_video/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Improve AEC controller time precision
+
 ## 1.2.0
 
 - Add multiple camera web server